fix: Entities with @Getter
Former-commit-id: a7bcf1ffc29d055959ec71c59c94addda47b86e2

+++ ondo/src/main/java/com/nextlevel/ondo/repository/FeedRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 // DAO
 // 자동으로 bean등록이 된다.
 // @Repository // 생략 가능하다.
-public interface FeedRepository extends JpaRepository<Feed, Integer> {
+public interface FeedRepository extends JpaRepository<Feed, Long> {
     Optional<Feed> findByFeedId(Long feedId);
     List<Feed> findAll();
 }
